+++ changelog.md
@@ -1,6 +1,18 @@
 # Changelog
 Service for creating and utilizing test data pools via REST API.
 
+## [1.7.4]
+
+### Changed
+
+- Mute deadlock when happens DB and application cache collision
+
+## [1.7.3]
+
+### Changed
+
+- Fix incorrect load big locked datapool
+
 ## [1.7.2]
 
 ### Changed